+++ spec/app/app/serializers/ams/custom_types_serializer.rb
@@ -30,7 +30,7 @@ module Ams
     typelize kind: "'user' | null"
     attribute :kind
 
-    typelize ["string", "number"]
+    typelize [:string, :number]
     attribute :tag
   end
 end