+++ spec/app/app/serializers/panko/custom_types_serializer.rb
@@ -30,7 +30,7 @@ module Panko
     typelize kind: "'user' | null"
     attributes :kind
 
-    typelize tag: ["string", "number"]
+    typelize tag: [:string, :number]
     attributes :tag
   end
 end